+++ Tests/XcodeTests/UIKitProject/UIKitProject/XibViewController3.swift
@@ -1,6 +1,11 @@
 import UIKit
 
 class XibViewController3: UIViewController {
-  @IBOutlet weak var button: UIButton!
-  @IBAction func clickFromSubclass(_ sender: Any) {}
+    // Referenced via XIB (connected)
+    @IBOutlet weak var button: UIButton!
+    @IBAction func click(_ sender: Any) {}
+
+    // Unreferenced - not connected in XIB
+    @IBOutlet weak var unusedOutlet: UILabel!
+    @IBAction func unusedAction(_ sender: Any) {}
 }